+++ viteclientts/src/main.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import './utils/axiosSetup.tsx';
 import App from './App.tsx'
 import './index.css'
 import './utils/i18n.tsx'